git-svn-id: http://multiwii.googlecode.com/svn/trunk/MultiWii_shared@650 02679b44-d973-9852-f2fa-63770883b36c

+++ RX.pde
@@ -343,7 +343,7 @@ uint16_t readRawRC(uint8_t chan) {
     if (rcFrameComplete) {
       for (uint8_t b = 3; b < SPEK_FRAME_SIZE; b += 2) {
         uint8_t spekChannel = 0x0F & (spekFrame[b - 1] >> SPEK_CHAN_SHIFT);
-        if (spekChannel < SPEK_MAX_CHANNEL) spekChannelData[spekChannel] = (long(spekFrame[b - 1] & SPEK_CHAN_MASK) << 8) + spekFrame[b];
+        if (spekChannel < SPEK_MAX_CHANNEL) spekChannelData[spekChannel] = ((uint32_t)(spekFrame[b - 1] & SPEK_CHAN_MASK) << 8) + spekFrame[b];
       }
       rcFrameComplete = 0;
     }
